Apply PSR-4 autoloader to tests folder
