+++ composer.json
@@ -20,11 +20,16 @@
     },
     "require-dev": {
         "twig/twig": "1.*",
-        "phpunit/phpunit": "^7.4"
+        "phpunit/phpunit": "^5.0"
     },
     "autoload": {
         "psr-0": {
             "Dflydev\\ApacheMimeTypes": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Dflydev\\ApacheMimeTypes\\Test\\": "tests"
         }
     },
     "extra": {
