+++ phpunit.xml.dist
@@ -2,7 +2,7 @@
 <phpunit colors="true" bootstrap="vendor/autoload.php">
     <testsuites>
         <testsuite name="Apache MIME Types Test Suite">
-            <directory>./tests/Dflydev/ApacheMimeTypes</directory>
+            <directory suffix="Test.php">./tests</directory>
         </testsuite>
     </testsuites>
