+++ tests/ArrayRepositoryTest.php
@@ -9,22 +9,27 @@
  * file that was distributed with this source code.
  */
 
-namespace Dflydev\ApacheMimeTypes;
+namespace Dflydev\ApacheMimeTypes\Test;
+
+use Dflydev\ApacheMimeTypes\ArrayRepository;
 
 /**
- * Flat Repository Test
+ * Array Repository Test
  *
  * @author Beau Simensen <beau@dflydev.com>
  */
-class FlatRepositoryTest extends AbstractRepositoryTest
+class ArrayRepositoryTest extends AbstractRepositoryTest
 {
     protected function createDefaultRepository()
     {
-        return new FlatRepository;
+        return null;
     }
 
     protected function createRepository()
     {
-        return new FlatRepository(__DIR__.'/Fixtures/mime.types');
+        return new ArrayRepository(array(
+            'dflydev/apache-mime-types' => array('dflydevamt', 'ddevamt'),
+            'dflydev/yet-another-mime-type' => array('dflydevyamt'),
+        ));
     }
 }
