+++ tests/JsonRepositoryTest.php
@@ -9,22 +9,24 @@
  * file that was distributed with this source code.
  */
 
-namespace Dflydev\ApacheMimeTypes;
+namespace Dflydev\ApacheMimeTypes\Test;
+
+use Dflydev\ApacheMimeTypes\JsonRepository;
 
 /**
  * JSON Repository Test
  *
  * @author Beau Simensen <beau@dflydev.com>
  */
-class PhpRepositoryTest extends AbstractRepositoryTest
+class JsonRepositoryTest extends AbstractRepositoryTest
 {
     protected function createDefaultRepository()
     {
-        return new PhpRepository;
+        return new JsonRepository;
     }
 
     protected function createRepository()
     {
-        return null;
+        return new JsonRepository(__DIR__.'/fixtures/mime.types.json');
     }
 }
