+++ tests/ParserTest.php
@@ -9,7 +9,9 @@
  * file that was distributed with this source code.
  */
 
-namespace Dflydev\ApacheMimeTypes;
+namespace Dflydev\ApacheMimeTypes\Test;
+
+use Dflydev\ApacheMimeTypes\Parser;
 
 /**
  * Parser Test
@@ -22,7 +24,7 @@ class ParserTest extends \PHPUnit_Framework_TestCase
     {
         $parser = new Parser;
 
-        $this->map = $parser->parse(__DIR__.'/../../../src/Dflydev/ApacheMimeTypes/Resources/mime.types');
+        $this->map = $parser->parse(__DIR__.'/../src/Dflydev/ApacheMimeTypes/Resources/mime.types');
     }
 
     public function testCss()
