+++ tests/fixtures/mime.types
